requirements.txt
setuptools>=65.0.0
wheel>=0.38.0
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
deep-sort-realtime>=1.3.0 